ListView: convert an if-else-if block into a nested-if
The fact that we had a comment "// model is null" in the else-if
part of this if-block was a strong evidence that the code was
not very readable to begin with, so we morph this to convert it
into a nested if block in the UpdateAdjustments() method, and
we remove the comment (even though this provides a higher level
of indentation, it will help to make the next commit after this,
much more understandable).

No change of behaviour in this commit.