ListView: zero vadjustment values also when model.Count is zero
One of the edge cases fixed by this recent commit [1] consisted of
zeroing all values of the vadjustment object when there were no rows
to show (model == null).

But, as evidenced when adding and subsequently removing tracks from
Banshee's PlayQueueSource as an example, there are obvious cases in
which model can be not null and still not hold anything to show in
the widget (model.Count == 0).

So the block where we zero all VAdjustment values (Upper, Lower,
PageSize, PageIncrement, StepIncrement, Value) needs to cover also
the case when model.Count == 0, otherwise there would be still some
rendering issues (SizeAllocated calls with negative coordinates).

We also add a TODO comment in this commit, to remind us in the future
that we could use Linq in the model to be a bit more efficient here.

[1] https://git.gnome.org/browse/hyena/commit/?h=gtk3&id=0745bfb75809886925dfa49a57c79e5f71565d08