ListView: fix rendering glitch when child_size changes (bgo#709523)
On using the ListView in non-ViewLayout mode (the deprecated way to use
this widget), there was a rendering glitch that was making gtk+ not
refresh its drawing when child size changed (i.e.: when new elements
were added on the model).

The way to fix it is calling QueueResize() when the measuring of the
child size happens, and this size ends up being different than the
previous time (if we did in all cases, the performance would be degraded
without need).