Log: make Exception() be higher severity (Error) if no msg is supplied
There are many places in apps in which Hyena.Log.Exception(ex)
is used to prevent the app from crashing, so it is an unexpected
exception, which should be raised as an error, not as a warning.

If a message is supplied though, it could mean that the exception
is expected to happen in an unlikely scenario, but not a bug.