Gui: avoid spinbuttons rendering out-of-screen (bgo#723811)
gtk3 has a bug (or a feature?) about rendering spinbuttons with
a very high maximum (i.e., Double.Max in .NET is 1.79769313486232E+308)
out of screen (instead of being horizontally-scrollable with the
keyboard like in gtk2). This has been reported as bgo#727294, but
using a workaround here is not a big deal for most usages (i.e not
a big deal that Banshee's Smart Playlists max limit are lowered to
a ceiling of 2,147,483,647 tracks).

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>