ListView: Rework rendering of header to follow GTK theme (bgo#709619)
We now use the StyleContext to draw the frame and background of the
header cells. By setting the correct region and flags, the result should
follow the GTK+ theme, giving us the same appearance a a standard
treeview.

But themes do not always account for custom widgets in their CSS
definitions. For example, with Adwaita the last header is not supposed
to have a right border, but for the ListView it still does, because of
an incorrect CSS selector (see bgo#731463).