Hyena.Gui: Fix handling of StateFlags when rendering custom widgets
GTK+ 3.x has added several flags to the StateFlags enum, and some of
them after 3.0. Of particular interest is the "backdrop" flag which
indicated that the window does not have the focus, and is heavily used
by the Adwaita theme to "tone down" stuff in the background.

This means we shoud never blindly overwrite a StateFlag value. So we
need to merge the StateFlags we want to set with the ones currently
set. For example, the result then follows both the cell state, like
"selected", and the overall widget state, like "backdrop".

We also to make sure that, after any modification to StyleContext.State,
the previous value is restored, so that we don't mess up the rest of
the rendering.