ListView: Refactor row selection rendering to use StyleContext.Render*
Use StyleContext.RenderBackground and RenderFrame with the proper class
and state to render selected rows.

This allows us to follow the GTK theme for selected rows, in particular
in the "backdrop" state, while getting rid of complicated Cairo code.

This removes the workaround for buggy themes, as we're not choosing the
color ourselves. But it should be OK, as the main culprit (Ambiance) was
fixed, and any similar problem that would now affect us would also
affect any custom treeview-like widget. So fix your themes !