ListView: Fix column headers appearance during drag-and-drop
Just add the "dnd" class on the column header, and remove all the custom
drawing we were doing in that situation.

Sadly, this class is not applied with the Adwaita theme (see
bgo#731663), so we don't get the background gradient that's supposed
to be there. But I think it looks good enough, and better than what we
had before.