Log: deprecate Exception() methods
As a follow-up to this improvement [1], it's actually better
to require the caller to be explicit about if she considers
the exception to be an error or a warning, so let's deprecate
the methods Exception(e) and Exception(msg,e) entirely and
recommend new overloads for Error() and Warning() that can
receive an Exception parameter.

The motivation for this is that I keep seeing many banshee
contributors use the Error() methods along with a message
simply extracted from the exception via 'ex.Message'. This,
in most cases, is unacceptable because it doesn't log the
stacktrace of the exception and doesn't aid debugging.

[1] https://git.gnome.org/browse/hyena/commit/?id=3feafb7328154b3c4901e8e57b70b05d9b4c9105