build: enable warnings as errors
Like it used to happen in banshee (in the stable branch it still
happens), let's enable warning as errors when ENABLE_RELEASE is
false. This allows us to not fall behind with warnings by mistake,
especially when we want to enable back warnings as errors again soon
in banshee.

For this, I fixed the remaining warnings resulting from the
deprecation of Log.Exception().