Windows: Specify CallingConvention everywhere (bgo#751045)
The default CallingConvention is WinAPI, which defaults to
the system's convention which might either be StdCall or CDecl.
For C libraries the CallingConvention to use is CDecl
else strange runtime errors might occur because of stack
corruption.

Since we are using C libraries only, specify the CDecl
CallingConvention everywhere.

On Windows the MDA in Visual Studio would detect a wrong
calling convention (mono has no such detection).