list: Move OnStyleUpdated from constructor
This commit moves the OnStyleUpdated call from the constructor to
immediately after the base.OnRealized call completes.

The logic in that call is shown to correct the view layout's squashed
rendering and inability to scroll when the window is resized for the
first time but it wasn't having the desired effect at construction.