H.Gui: ListView, respect fixed sized columns
This commit supplements the measuring logic for text rows by selecting
all the fixed size column cells and choosing the maximum width and
height from each.

This allows fancy columns that take more height to render correctly and
is accompanied by a small change in header handling as we invalidate
measurements when a column is added or removed or the column controller
is changed.